refactor: error dialog
